build: Makefile now works from within GOPATH
Previously, the Makefile would always create a symbolic link from the current
working directory to the correct place in the GOPATH. That way, running "make"
from the directory in which your code was checked out would "just work."

However, if your current working directory _was_ in the GOPATH, this link is
unnecessary, and in fact will cause an error. This modified Makefile skips
creating the link unnecessarily.